requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
Pillow==5.4.1
PyAudio==0.2.11
requests==2.21.0
SpeechRecognition==3.8.1
urllib3==1.24.1
